Move completions it seperate file
